Declare that I am pausing until I have a team
This is to all of you guys, I am pausing due to the lack of people (just me), working on it, I feel a bit stressed on it and I will pause development on the mod until I can assemble a proper team to work on the mod, sorry for the inconvience and thank you for your paitence. I hope to see you as a team than as a single person.